feat(admin-navigation.tsx): add user id to user and settings navigation links feat(routes.tsx): add dynamic user id to user and settings routes feat(edit-user): add toast notification on successful user update fix(edit-user): fix backAction url to include dynamic user id refactor(edit-user): remove unused action and loader files from user and settings pages fix(action.ts): remove redirect import from action function
