--- src/pages/admin/my/settings/action.ts
@@ -1,2 +1,0 @@
-import { action } from "../user/action";
-export { action };
