+++ src/pages/admin/my/settings/index.ts
@@ -1,3 +1,4 @@
-export * from "./action";
+import { action, loader } from "~/pages/admin/users/edit-user";
+export { action, loader };
+
 export * from "./component";
-export * from "./loader";
