--- src/pages/admin/my/settings/loader.ts
@@ -1,2 +1,0 @@
-import { loader } from "../user/loader";
-export { loader };
