+++ src/pages/admin/my/user/component.tsx
@@ -3,11 +3,10 @@ import { useEffect } from "react";
 import { useActionData, useLoaderData } from "react-router-dom";
 import { BadgeStatus } from "~/components/badge-status";
 import { UserForm } from "~/components/user/user-form";
+import { action, loader } from "~/pages/admin/users/edit-user";
 import { useToast } from "~/providers/toast";
 import { useTranslation } from "~/providers/translate-provider";
 import { isUser } from "~/types/user";
-import { action } from "./action";
-import { loader } from "./loader";
 
 export function Component() {
   const { t } = useTranslation({ id: "my-user", locales });
