--- src/pages/admin/my/user/loader.ts
@@ -1,11 +1,0 @@
-import { getUserGetByIdQueryOptions } from "~/api/bookingShopifyApi";
-import { getToken } from "~/providers/ability-provider";
-import { queryClient } from "~/providers/query-provider";
-
-export const loader = async () => {
-  const user = getToken();
-  const response = await queryClient.fetchQuery(
-    getUserGetByIdQueryOptions(user.userId)
-  );
-  return response.data.payload;
-};
